[ge-learning] Change distribution mutation to mutate one divider
This will make the SA algorithm search in a spaces with one dimension less,
wich could increase performance. The space still covers all values,
but dosn't visit values that have the same meaning (e.g. [0.5, 0.5] has same meaning as [1.0, 1.0]).
The dimension reduction is done by converting the set of N weights to a set of N-1 dividers.
After mutation, it can be converted back to weights.

Additionally, it now only mutates one divider instead of all weights.
This is to improve locality between neighbours, and because changing dividers has
a bigger effect on the distribution than changing weights.