[ge-learning] Change weight/divider conversion to a simpler test-passing method
Now each divider is the proportion between the sum of the weights before the divder
divided on the same sum plus the next weight.

This way, any NaN and inconsistency issues should be resolved.
The drawback may be that modifications to dividers may not have a uniform change on the weights.