[dgel-ge] Change Simulator to match paper and allow custom operators
It now matches the description in 'A Comparison of Selection Schemes
used in Genetic Algorithms', Tobias Blickle and Lothar Thiele, 1995.
This will make it easier to use the paper as a source for decisions, and
it might perform better?

It also allows for and requires custom operators, such that other
operators than crossover and mutation may be implemented.